pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-demo"
version = "0.1.0"
description = "Evolution of RAG from vanilla to Graph-RAG implementations with educational examples"
readme = "README.md"
license = "MIT"
authors = [{name = "Karthik Ravi"}]
requires-python = ">=3.11"
keywords = ["rag", "retrieval-augmented-generation", "graph-rag", "nlp", "ai", "demo"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.5.0",
    "plotly>=5.10.0",
    "nltk>=3.6",
    "tiktoken>=0.4.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.20.0",
    "openai>=1.0.0",
    "networkx>=2.8.0",
    "pillow>=8.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
    "black",
    "pytest>=7.0.0",
    "pytest-asyncio",
]

[project.urls]
Homepage = "https://github.com/karthik-ravi-1537/RAG-Demo"
Repository = "https://github.com/karthik-ravi-1537/RAG-Demo.git"
Issues = "https://github.com/karthik-ravi-1537/RAG-Demo/issues"

[tool.black]
line-length = 120
skip-string-normalization = false
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = [".venv", "venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes errors
    "I",    # import sorting
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function calls in argument defaults
    "B904",  # exception chaining (acceptable for demo project)
    "E402",  # module level import not at top (acceptable in test files)
    "N818",  # exception naming (acceptable for demo)
]

[tool.ruff.lint.isort]
force-single-line = false
force-sort-within-sections = false

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
